Implemented functionality to add and remove points.

+++ app/src/main/java/com/tinytotsnbites/papapoints/data/DAOs.kt
@@ -69,5 +69,5 @@ interface RatingDao {
     fun getByTaskId(taskId: Long): List<Rating>
 
     @Query("SELECT SUM(rating) FROM rating")
-    fun getTotalRating(): Float
+    fun getTotalRating(): Int
 }
